(PC-5532) move files
