+++ src/features/navigation/mocks/navigationStateSnapshots.ts
